added dept_class.py
printed the list of roll_nos that are enrolled for the courses.
printed all the courses student is taking.
verified if the professor is HoD or not.
printed all the courses in department.
printed all the professors in department.
according to semester and year, answered if the course is offered or not
to respective sem or year.